[Tests] Move PyRetainingSharedPtr test bindings
In preparation for adding more tests in #508, move the bindings
specific to the `PyRetainingSharedPtr` tests to their own cpp file.
